FrictionalSliding constraints resist non-normal motion.
In this commit, the "sliding" portion of the frictional sliding
constraint generates a force to try and keep the suture's
localization from moving away from the tissue's localization,
regardless of the direction of motion.
This change means that a suture sliding through the point at an
angle will have a correcting force pointed along the direction of
motion. Previously the friction force solely acted
normal to the triangle.