Adjusts projections in KnotIDBehavior.
The previous approach of using makeRigidTransform was confusing
because the vector arguments to the function can be confused for
the resulting matrix.  With operator<<() it is clear that these three
projection matrices are orthogonal.

Also, adds a render test of a square knot that fails because
the non-enhanced gauss code cannot distinguish between a
granny knot and a square knot in all orienations.