KnotIdentificationBehavior uses enhanced gauss codes.
The enhanced gauss code additionally incorporates the handedness of each
crossing, i.e., the direction of the over-strand with respect to the
direction of the under-strand.
Without this additional information, some
views of our sample square knot would result in the correct calculation
of the gauss code we expected for a granny knot.  See the
SquareKnotRotated render test in the previous commit (though an
orthogonal render view would be necessary to match the info used
by the ID algorithm). That issue is a limitation with the non-enhanced
gauss code, and is addressed with the enhanced gauss code used here.

Two known issues remain.  First, some projections generate complex
(enhanced) gauss codes that we are not able to simplify to the expected
gauss code, so the behavior labels it an unknown knot. Second, some
projections generate simpler (enhanced) gauss codes, e.g., a granny knot
is identified as trefoil.  Possibly in those cases, the projection is
parallel to one of the suture ends in a way that reduces the number of
crossings.  We may get better results by forcing the behavior to run all
three projections and then vote.