KnotIdentificationBehavior always runs three projections for robustness.
Depending on the orientation of the knot in space, one or more
projections can result in non-ideal results from the identification
algorithm.  Specifically, one of our sample knots with just the
right orientation will get an enhanced gauss code that's too long
so it's labeled an unknown knot.  Or occasionally the gauss code
is too short and a granny knot is identified as trefoil.

I do not know how these misidentifications could be prevented, but by
forcing the algorithm to always run on all three projections, our random
orientation testing succeeds more often, though it still fails
occasionally.