KnotIdentificationBehavior closes the loop.
To prevent an orientation directly along the free end of the knot,
which could reduce the number of crossings and prevent identification
of a knot, we add a segment from the last node to the first node.

This approach mimics theoretical knot theory in which all knots
are loops with no free ends.

On this commit, the random orientation test runs for longer
before failing.

A potential downside is if the segment between the two ends
would pass through the knot, causing unwanted crossings.