KnotIdentificationBehavior uses six projections.
By adding three 45-degree off-axis projections, the results are
more accurate.  Previously, there were some test cases in which the
projections would result in one unknown knot, one correct,
and one incorrect (e.g., trefoil instead of granny). With these
six projections, there seem to be enough correct results that
the correct knot is identified instead of an incorrect knot.
The random test ran without failure for 28,000 orientations before
I stopped it.

This commit will double the time to update the behavior.
If update rate is critical, it would be possible to do an early return
in some cases, or possibly use only 4 or 5 projections.