KnotIdentificationBehaviorTests test the ReidMeister moves again.
The tests had been commented out due to interface changes.

The unit tests for variable start crossings has been removed, because
the behavior no longer recognizes knots that start anywhere but
the beginning of the gauss code.  That concept may make sense for
theoretical knots that are loops without a clear start location,
but in our practical simulations the knot always starts from a
free end of the suture.